# docs/swagger.yaml
basePath: /
definitions:
...
  resource.InputSocialMedia:
    properties:
      name:
        example: Instagram
        type: string
      social_media_url:
        example: www.instagram.com/uzumakinaruto
        type: string
    required:
    - name
    - social_media_url
    type: object
...
info:
  contact:
    email: kadekbintanga@gmail.com
    name: API Support
  description: This is a service for final project
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: Final Project by Bintang
  version: "1.0"
paths:
...
  /api/v1/socialmedia:
    get:
      description: Get social media with bearer token
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user social media
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Create a social media with the input payload
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputSocialMedia'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a social media
      tags:
      - Social Media
  /api/v1/socialmedia/{socialmedia_id}:
    delete:
      consumes:
      - application/json
      description: delete a social media with the token
      parameters:
      - description: SocialMedia ID
        in: path
        name: socialmedia_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a social media
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update a social media with the input payload
      parameters:
      - description: SocialMedia ID
        in: path
        name: socialmedia_id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputSocialMedia'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a social media
      tags:
      - Social Media
...
swagger: "2.0"